Reorganiser compiler in terms of optimisers and transformers
Passes that are required for a circuit to usable, rather than just for improving performance, are now called transformers instead of optimisations.

Transforms:
fallback
r1cs
csat

Optimisations:
general
more to come

General optimising is now separated into its own pass instead of being executed by r1cs of csat. As a result the r1cs pass current has no work to perform.
