+++ acir/src/native_types/witness.rs
@@ -6,6 +6,7 @@ use flate2::{
 };
 use serde::{Deserialize, Serialize};
 
+// Witness might be a misnomer. This is an index that represents the position a witness will take
 #[derive(
     Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Default, Serialize, Deserialize,
 )]
