+++ acvm/src/compiler.rs
@@ -1,6 +1,6 @@
 // The various passes that we can use over ACIR
-pub mod fallback;
-pub mod optimiser;
+pub mod optimisers;
+pub mod transformers;
 
 use crate::Language;
 use acir::{
@@ -9,10 +9,9 @@ use acir::{
     BlackBoxFunc,
 };
 use indexmap::IndexMap;
-use optimiser::{CSatOptimiser, GeneralOptimiser};
+use optimisers::GeneralOptimiser;
 use thiserror::Error;
-
-use self::{fallback::IsBlackBoxSupported, optimiser::R1CSOptimiser};
+use transformers::{CSatTransformer, FallbackTransformer, IsBlackBoxSupported, R1CSTransformer};
 
 #[derive(PartialEq, Eq, Debug, Error)]
 pub enum CompileError {
@@ -29,33 +28,48 @@ pub fn compile(
     // Currently the optimiser and reducer are one in the same
     // for CSAT
 
-    // Fallback pass
-    let fallback = fallback::fallback(acir, is_blackbox_supported)?;
+    // Fallback transformer pass
+    let acir = FallbackTransformer::transform(acir, is_blackbox_supported)?;
 
-    let optimiser = match &np_language {
+    // General optimiser pass
+    let mut opcodes: Vec<Opcode> = Vec::new();
+    for opcode in acir.opcodes {
+        match opcode {
+            Opcode::Arithmetic(arith_expr) => {
+                opcodes.push(Opcode::Arithmetic(GeneralOptimiser::optimise(arith_expr)))
+            }
+            other_gate => opcodes.push(other_gate),
+        };
+    }
+    let acir = Circuit { opcodes, ..acir };
+
+    let transformer = match &np_language {
         crate::Language::R1CS => {
-            let optimiser = R1CSOptimiser::new(fallback);
-            return Ok(optimiser.optimise());
+            let transformer = R1CSTransformer::new(acir);
+            return Ok(transformer.transform());
         }
-        crate::Language::PLONKCSat { width } => CSatOptimiser::new(*width),
+        crate::Language::PLONKCSat { width } => CSatTransformer::new(*width),
     };
 
-    // TODO: the code below is only for CSAT optimiser
+    // TODO: the code below is only for CSAT transformer
     // TODO it may be possible to refactor it in a way that we do not need to return early from the r1cs
     // TODO or at the very least, we could put all of it inside of CSATOptimiser pass
 
     // Optimise the arithmetic gates by reducing them into the correct width and
     // creating intermediate variables when necessary
-    let mut optimised_gates = Vec::new();
+    let mut transformed_gates = Vec::new();
 
-    let mut next_witness_index = fallback.current_witness_index + 1;
-    for opcode in fallback.opcodes {
+    let mut next_witness_index = acir.current_witness_index + 1;
+    for opcode in acir.opcodes {
         match opcode {
             Opcode::Arithmetic(arith_expr) => {
                 let mut intermediate_variables: IndexMap<Witness, Expression> = IndexMap::new();
 
-                let arith_expr =
-                    optimiser.optimise(arith_expr, &mut intermediate_variables, next_witness_index);
+                let arith_expr = transformer.transform(
+                    arith_expr,
+                    &mut intermediate_variables,
+                    next_witness_index,
+                );
 
                 // Update next_witness counter
                 next_witness_index += intermediate_variables.len() as u32;
@@ -67,10 +81,10 @@ pub fn compile(
                 new_gates.push(arith_expr);
                 new_gates.sort();
                 for gate in new_gates {
-                    optimised_gates.push(Opcode::Arithmetic(gate));
+                    transformed_gates.push(Opcode::Arithmetic(gate));
                 }
             }
-            other_gate => optimised_gates.push(other_gate),
+            other_gate => transformed_gates.push(other_gate),
         }
     }
 
@@ -78,7 +92,7 @@ pub fn compile(
 
     Ok(Circuit {
         current_witness_index,
-        opcodes: optimised_gates,
-        public_inputs: fallback.public_inputs, // The optimiser does not add public inputs
+        opcodes: transformed_gates,
+        public_inputs: acir.public_inputs, // The optimiser does not add public inputs
     })
 }
