+++ acvm/src/compiler/optimisers/mod.rs
@@ -1,0 +1,3 @@
+mod general;
+
+pub use general::GeneralOpt as GeneralOptimiser;
