+++ acvm/src/compiler/transformers/r1cs.rs
@@ -1,0 +1,15 @@
+use acir::circuit::Circuit;
+
+pub struct R1CSTransformer {
+    acir: Circuit,
+}
+
+impl R1CSTransformer {
+    pub fn new(acir: Circuit) -> Self {
+        Self { acir }
+    }
+    // TODO: We could possibly make sure that all polynomials are at most degree-2
+    pub fn transform(self) -> Circuit {
+        self.acir
+    }
+}
